Question to ask why the event is not bubbling
My understanding is that the default of bubbling is true. So when
mouse is over a box (one single grid box) and background color
is changed to lightpink, its container (.container)
will also have turn into lightpink.

But reality is only the single box container's background color is
changed, not its parent.